+++ src/Core/Exception/UploadFileException.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace AppTank\Horus\Core\Exception;
+
+/**
+ * @internal Class UploadFileException
+ *
+ * Exception thrown when an error occurs while uploading a file.
+ *
+ * @package AppTank\Horus\Core\Exception
+ *
+ * @author John Ospina
+ * Year: 2024
+ */
+class UploadFileException extends \Exception
+{
+
+}