+++ tests/Feature/Api/GetWrapperFileApiTest.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace Api;
+
+use AppTank\Horus\Core\File\IFileHandler;
+use AppTank\Horus\RouteName;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use Mockery\Mock;
+use Tests\_Stubs\SyncFileUploadedModelFactory;
+use Tests\Feature\Api\ApiTestCase;
+
+class GetWrapperFileApiTest extends ApiTestCase
+{
+
+    use RefreshDatabase;
+
+
+    private IFileHandler|Mock $fileHandler;
+
+    function setUp(): void
+    {
+        parent::setUp();
+
+        $this->fileHandler = $this->mock(IFileHandler::class);
+        $this->app->bind(IFileHandler::class, fn() => $this->fileHandler);
+    }
+
+    function testGetFileImageIsSuccess()
+    {
+        // Given
+        $imageUrl = "https://images.unsplash.com/photo-1471502041392-3182a8a099a5?q=80&w=1740&auto=format&fit=crop";
+        $fileUploaded = SyncFileUploadedModelFactory::create(fileUrl: $imageUrl);
+
+        // When
+        $response = $this->getJson(route(RouteName::GET_WRAPPER_FILE->value, ["id" => $fileUploaded->getId()]));
+
+        // Then
+        $response->assertOk();
+        $response->assertHeader("Content-Type", "image/jpeg");
+    }
+
+    function testGetFilePdfIsSuccess()
+    {
+        // Given
+        $pdfUrl = "https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy.pdf";
+        $fileUploaded = SyncFileUploadedModelFactory::create(fileUrl: $pdfUrl);
+
+        // When
+        $response = $this->getJson(route(RouteName::GET_WRAPPER_FILE->value, ["id" => $fileUploaded->getId()]));
+
+        // Then
+        $response->assertOk();
+        $response->assertHeader("Content-Type", "application/pdf; qs=0.001");
+    }
+
+    function testGetFileIsNotFound()
+    {
+        // Given
+        $fileId = $this->faker->uuid;
+
+        // When
+        $response = $this->getJson(route(RouteName::GET_WRAPPER_FILE->value, ["id" => $fileId]));
+
+        // Then
+        $response->assertNotFound();
+    }
+}