+++ tests/Unit/Repository/EloquentEntityRepositoryTest.php
@@ -3,6 +3,7 @@
 namespace Tests\Unit\Repository;
 
 
+use AppTank\Horus\Core\Entity\EntityReference;
 use AppTank\Horus\Core\Exception\OperationNotPermittedException;
 use AppTank\Horus\Core\Factory\EntityOperationFactory;
 use AppTank\Horus\Core\Hasher;
@@ -74,8 +75,8 @@ class EloquentEntityRepositoryTest extends TestCase
         $this->entityRepository->insert(...$operations);
 
         // Then
-         $this->assertDatabaseCount(ParentFakeWritableEntity::getTableName(), count($parentsEntities) + count($childEntities));
-         $this->assertDatabaseCount(ChildFakeWritableEntity::getTableName(), count($childEntities));
+        $this->assertDatabaseCount(ParentFakeWritableEntity::getTableName(), count($parentsEntities) + count($childEntities));
+        $this->assertDatabaseCount(ChildFakeWritableEntity::getTableName(), count($childEntities));
 
         foreach ($parentsEntities as $entity) {
             $expectedData = $entity->toArray();
@@ -153,7 +154,8 @@ class EloquentEntityRepositoryTest extends TestCase
             $expectedData = array_merge(
                 ["id" => $operation->id,
                     ParentFakeWritableEntity::ATTR_NAME => $parentsEntities[$index]->name,
-                    ParentFakeWritableEntity::ATTR_ENUM => $parentsEntities[$index]->value_enum
+                    ParentFakeWritableEntity::ATTR_ENUM => $parentsEntities[$index]->value_enum,
+                    ParentFakeWritableEntity::ATTR_IMAGE => $parentsEntities[$index]->image
                 ],
                 $operation->attributes
             );
@@ -193,7 +195,9 @@ class EloquentEntityRepositoryTest extends TestCase
         $expectedData = array_merge(
             ["id" => $lastOperation->id,
                 ParentFakeWritableEntity::ATTR_NAME => $parentEntity->name,
-                ParentFakeWritableEntity::ATTR_ENUM => $parentEntity->value_enum],
+                ParentFakeWritableEntity::ATTR_ENUM => $parentEntity->value_enum,
+                ParentFakeWritableEntity::ATTR_IMAGE => $parentEntity->image
+            ],
             $lastOperation->attributes
         );
 
@@ -613,5 +617,23 @@ class EloquentEntityRepositoryTest extends TestCase
 
         // Then
         $this->assertFalse($result);
+    }
+
+    function testGetEntityPathHierarchy()
+    {
+        $userOwnerId = $this->faker->uuid;
+
+        $parentEntity = ParentFakeEntityFactory::create($userOwnerId);
+        $childEntity = ChildFakeEntityFactory::create($parentEntity->getId(), $userOwnerId);
+
+        $entityReference = new EntityReference(ChildFakeWritableEntity::getEntityName(), $childEntity->getId());
+
+        // When
+        $result = $this->entityRepository->getEntityPathHierarchy($entityReference);
+
+        // Then
+        $this->assertCount(2, $result);
+        $this->assertEquals($parentEntity->getId(), $result[0]->getId());
+        $this->assertEquals($childEntity->getId(), $result[1]->getId());
     }
 }